Added leader activity and various resources from the website
Created a new leader activity with a spinner to query the server for
leader stats (not implemented yet, nor is server selection on this
page).  Added resources from online (pngs mostly), scaled down the
background image, as before it was too large for the screen and could
not be loaded.